move server render to separate location
