+++ app/server.js
@@ -1,21 +1,20 @@
 import Express from 'express';
 import React from 'react';
 import ReactDOM from 'react-dom/server';
-import { ApolloProvider } from 'react-apollo';
 import favicon from 'serve-favicon';
 import compression from 'compression';
 import httpProxy from 'http-proxy';
 import path from 'path';
 import PrettyError from 'pretty-error';
 import http from 'http';
-import _ from 'lodash';
 
-import { match, RouterContext } from 'react-router';
+import { match } from 'react-router';
 
 import config from '../config';
 import Html from './Html';
 import ourCreateClient from './apollo-client';
 import ourCreateStore from './store';
+import defaultSSR from './server-render/apollo-ssr';
 
 const targetUrl = 'http://' + config.apiHost + ':' + config.apiPort;
 const pretty = new PrettyError();
@@ -23,6 +22,7 @@ const pretty = new PrettyError();
 export default ({
   routes,
   disableSSR = false,
+  serverRender = defaultSSR,
   createClient = ourCreateClient,
   createStore = ourCreateStore }) => {
   const app = new Express();
@@ -65,7 +65,7 @@ export default ({
 
   app.use((req, res) => {
     if (__DEVELOPMENT__) {
-      // Do not cache webpack stats: the script file would change since
+      // Do not cache webpaRouterContextck stats: the script file would change since
       // hot module replacement is enabled in the development env
       webpackIsomorphicTools.refresh();
     }
@@ -87,41 +87,14 @@ export default ({
         res.status(500);
         hydrateOnClient();
       } else if (renderProps) {
-        // transfer request headers to networkInterface so that they're accessible to proxy server
-        // Addresses this issue: https://github.com/matthew-andrews/isomorphic-fetch/issues/83
-        const client = createClient(req.headers);
+        const client = createClient();
         const store = createStore({ client });
-
-
         const component = (
           <ApolloProvider client={client} store={store}>
             <RouterContext {...renderProps} />
           </ApolloProvider>
         );
-
-        const maybeRenderPage = () => {
-          if (!_.some(store.getState().apollo.queries, 'loading')) {
-            stopSubscription();
-            res.status(200);
-
-            global.navigator = {userAgent: req.headers['user-agent']};
-
-            const html = (
-              <Html assets={webpackIsomorphicTools.assets()}
-                component={component} store={store} />
-            );
-            res.send('<!doctype html>\n' + ReactDOM.renderToStaticMarkup(html));
-          }
-        }
-
-        // now wait for all queries in the store to go ready
-        const stopSubscription = store.subscribe(maybeRenderPage);
-
-        // render once, to initialize apollo queries
-        ReactDOM.renderToString(component);
-
-        // if the page has no queries, the store will never change
-        maybeRenderPage();
+        serverRender({renderProps, client, store, Html, req, res, component});
       } else {
         res.status(404).send('Not found');
       }
